refactor: Move heartbeat utilities to specific modules
