+++ posthog/temporal/batch_exports/snowflake_batch_export.py
@@ -51,7 +51,7 @@ from posthog.temporal.batch_exports.utils import (
 from posthog.temporal.common.clickhouse import get_client
 from posthog.temporal.common.heartbeat import Heartbeater
 from posthog.temporal.common.logger import bind_temporal_worker_logger
-from posthog.temporal.common.utils import (
+from posthog.temporal.batch_exports.heartbeat import (
     BatchExportRangeHeartbeatDetails,
     DateRange,
     HeartbeatParseError,
@@ -589,7 +589,7 @@ async def insert_into_snowflake_activity(inputs: SnowflakeInsertInputs) -> Recor
         if not await client.is_alive():
             raise ConnectionError("Cannot establish connection to ClickHouse")
 
-        _, details = await should_resume_from_activity_heartbeat(activity, SnowflakeHeartbeatDetails, logger)
+        _, details = await should_resume_from_activity_heartbeat(activity, SnowflakeHeartbeatDetails)
         if details is None:
             details = SnowflakeHeartbeatDetails()
